+++ base/setup/lib/precomp.h
@@ -55,4 +55,4 @@
 extern HANDLE ProcessHeap;
 
 #include "errorcode.h"
-#include "linklist.h"
+#include "utils/linklist.h"
