+++ base/setup/lib/setuplib.h
@@ -26,20 +26,20 @@
 extern HANDLE ProcessHeap;
 
 #include "errorcode.h"
-#include "linklist.h"
-#include "ntverrsrc.h"
-// #include "arcname.h"
-#include "bldrsup.h"
+#include "utils/linklist.h"
+#include "utils/ntverrsrc.h"
+// #include "utils/arcname.h"
+#include "utils/bldrsup.h"
 #include "bootsup.h"
-#include "filesup.h"
+#include "utils/filesup.h"
 #include "fsutil.h"
-#include "genlist.h"
-#include "infsupp.h"
-#include "inicache.h"
-#include "partlist.h"
-#include "arcname.h"
-#include "osdetect.h"
-#include "regutil.h"
+#include "utils/genlist.h"
+#include "utils/infsupp.h"
+#include "utils/inicache.h"
+#include "utils/partlist.h"
+#include "utils/arcname.h"
+#include "utils/osdetect.h"
+#include "utils/regutil.h"
 #include "registry.h"
