+++ base/setup/lib/utils/linklist.h
