+++ base/setup/lib/utils/ntverrsrc.c
