+++ base/setup/lib/utils/ntverrsrc.h
