+++ base/setup/usetup/genlist.h
@@ -26,7 +26,7 @@
 
 #pragma once
 
-#include "../lib/genlist.h"
+// #include "../lib/utils/genlist.h"
 
 typedef struct _GENERIC_LIST_UI
 {
