+++ base/setup/usetup/partlist.h
@@ -25,7 +25,7 @@
 
 #pragma once
 
-#include "../lib/partlist.h"
+// #include "../lib/utils/partlist.h"
 
 typedef enum _FORMATMACHINESTATE
 {
